refactor(settings): 拆出通用状态/搜索/快捷键 hook 与侧边栏组件
- 新增 settings/：useSettingsGeneralState（终端点击行为/确认开关/工作区/GPU 状态）、useSettingsSearch（设置搜索与跳转高亮）、useSettingsShortcuts（快捷键录制与 Esc 关闭）、SettingsSidebar（侧边栏搜索与标签导航）
- SettingsModal.tsx 718 -> 361 行
